Fix data consistency: fractional lx support, MortalityTable XML drift
- 2010CM mortality lx values are fractional (e.g. 99382.28) but
  MortalityTableRow.Lx was int and tblMortality.lx was [int] - the C#
  pipeline could not read its own committed MortalityTable.json. Lx is
  now double and the DB column float (schema redeploy required)
- Regenerate XMLFiles/*.xml from JSONFiles/*.json via new JsonToXml.py
  so both source formats agree (MortalityTable.xml was missing the
  2010 year rows); regenerated XML verified round-trip identical
- Investigated TableJ ≡ TableK duplication: TableK values verified
  against the source PDF (450/450 rows for the 2.2-22.0% grid); TableJ
  is a byte-copy of TableK data although IRS Table J (annuities payable
  at the beginning of each interval) must differ - flagged in docs
- TableJ/K JSONs contain 50 rows (rates 0.2-2.0%) present in no 90CM
  source; provenance unknown, documented

+++ src/DataProcessingApp.Core/DataObjects/MortalityTableRow.cs
@@ -9,6 +9,6 @@
         public int Age { get; set; }
 
         [DbColumn("lx")]
-        public int Lx { get; set; }
+        public double Lx { get; set; }
     }
 }
